# Avoid importing pkg_resources