pkg_resources is expensive to load, which takes more than 400ms here.

Let's use importlib.metadata from stdlib for Python 3.8+ for entry points,
and packaging.version for version parsing instead. It boosts start up
time significantly.

Some benchmark data:

On `fuo --help`, which involves the version parsing but not the entry
points:

Before: 0.908s
After: 0.398s

On `fuo search test` with an added "sys.exit(1)" clause right after
scanning the entry points:

Before: 0.991s
After: 0.502s